probe/requirements.txt
kubernetes
pyyaml
requests
grpcio
grpcio-reflection
pika
confluent-kafka
openai
fastapi
uvicorn
httpx
